requirements.txt
numpy==1.19.2
torch==1.6.0
torchvision==0.7.0
PyYAML~=5.3.1